test(simulator): restructure postTransactionDelay tests per ModuleName-Feature naming convention
